requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
camel-converter==4.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
curl_cffi==0.11.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
executing==2.2.0
fonttools==4.58.0
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.25
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
meilisearch==0.34.1
multitasking==0.0.11
narwhals==1.39.1
nest-asyncio==1.6.0
numpy==2.2.5
openai==1.78.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.0
prompt_toolkit==3.0.51
protobuf==6.31.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
streamlit==1.45.1
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
yfinance==0.2.61
zstandard==0.23.0
